refactoring for unit testing
