+++ DirectoryServiceAPI/Services/Interfaces/IADFactory.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace DirectoryServiceAPI.Services.Interfaces
+namespace DirectoryServiceAPI.Services
 {
     public interface IADFactory //Creator
     {
